# packages/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Enspatium API",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "enspatium_session"
      }
    },
...
  },
  "paths": {
...
    "/namespaces/{namespaceSlug}/spaces": {
      "post": {
        "operationId": "createSpace",
        "tags": [
          "spaces"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "slug",
                  "type"
                ],
                "properties": {
                  "name": {
                    "minLength": 1,
                    "maxLength": 100,
                    "type": "string"
                  },
                  "slug": {
                    "minLength": 1,
                    "maxLength": 100,
                    "type": "string"
                  },
                  "type": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "git"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "object"
                        ]
                      }
                    ]
                  },
                  "visibility": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "public"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "private"
                        ]
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "type": "string"
            },
            "in": "path",
            "name": "namespaceSlug",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "namespaceId",
                    "createdByUserId",
                    "name",
                    "slug",
                    "type",
                    "visibility",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "namespaceId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "createdByUserId": {
                      "anyOf": [
                        {
                          "format": "uuid",
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "type": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "git"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "object"
                          ]
                        }
                      ]
                    },
                    "visibility": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "public"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "private"
                          ]
                        }
                      ]
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listSpaces",
        "tags": [
          "spaces"
        ],
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "type": "string"
            },
            "in": "path",
            "name": "namespaceSlug",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "namespaceId",
                      "createdByUserId",
                      "name",
                      "slug",
                      "type",
                      "visibility",
                      "createdAt",
                      "updatedAt"
                    ],
                    "properties": {
                      "id": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "namespaceId": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "createdByUserId": {
                        "anyOf": [
                          {
                            "format": "uuid",
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "name": {
                        "type": "string"
                      },
                      "slug": {
                        "type": "string"
                      },
                      "type": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "git"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "object"
                            ]
                          }
                        ]
                      },
                      "visibility": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "public"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "private"
                            ]
                          }
                        ]
                      },
                      "createdAt": {
                        "type": "string"
                      },
                      "updatedAt": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/namespaces/{namespaceSlug}/spaces/{spaceSlug}": {
      "get": {
        "operationId": "getSpace",
        "tags": [
          "spaces"
        ],
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "type": "string"
            },
            "in": "path",
            "name": "namespaceSlug",
            "required": true
          },
          {
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "type": "string"
            },
            "in": "path",
            "name": "spaceSlug",
            "required": true
          }
        ],
        "security": [
          {},
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "namespaceId",
                    "createdByUserId",
                    "name",
                    "slug",
                    "type",
                    "visibility",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "namespaceId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "createdByUserId": {
                      "anyOf": [
                        {
                          "format": "uuid",
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "type": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "git"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "object"
                          ]
                        }
                      ]
                    },
                    "visibility": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "public"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "private"
                          ]
                        }
                      ]
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateSpace",
        "tags": [
          "spaces"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "minProperties": 1,
                "type": "object",
                "properties": {
                  "name": {
                    "minLength": 1,
                    "maxLength": 100,
                    "type": "string"
                  },
                  "visibility": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "public"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "private"
                        ]
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "type": "string"
            },
            "in": "path",
            "name": "namespaceSlug",
            "required": true
          },
          {
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "type": "string"
            },
            "in": "path",
            "name": "spaceSlug",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "namespaceId",
                    "createdByUserId",
                    "name",
                    "slug",
                    "type",
                    "visibility",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "namespaceId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "createdByUserId": {
                      "anyOf": [
                        {
                          "format": "uuid",
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "type": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "git"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "object"
                          ]
                        }
                      ]
                    },
                    "visibility": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "public"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "private"
                          ]
                        }
                      ]
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteSpace",
        "tags": [
          "spaces"
        ],
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "type": "string"
            },
            "in": "path",
            "name": "namespaceSlug",
            "required": true
          },
          {
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "type": "string"
            },
            "in": "path",
            "name": "spaceSlug",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          }
        }
      }
    },
    "/namespaces/{namespaceSlug}/spaces/{spaceSlug}/git": {
      "get": {
        "operationId": "getGitSpaceInfo",
        "tags": [
          "spaces"
        ],
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "type": "string"
            },
            "in": "path",
            "name": "namespaceSlug",
            "required": true
          },
          {
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "type": "string"
            },
            "in": "path",
            "name": "spaceSlug",
            "required": true
          }
        ],
        "security": [
          {},
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "defaultBranch",
                    "branches",
                    "commits"
                  ],
                  "properties": {
                    "defaultBranch": {
                      "type": "string"
                    },
                    "branches": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "commits": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "shortId",
                          "authorName",
                          "authorEmail",
                          "authoredAt",
                          "message"
                        ],
                        "properties": {
                          "id": {
                            "pattern": "^[0-9a-f]{40,64}$",
                            "type": "string"
                          },
                          "shortId": {
                            "pattern": "^[0-9a-f]+$",
                            "type": "string"
                          },
                          "authorName": {
                            "type": "string"
                          },
                          "authorEmail": {
                            "type": "string"
                          },
                          "authoredAt": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/namespaces/{namespaceSlug}/spaces/{spaceSlug}/git/tags": {
      "get": {
        "operationId": "getGitSpaceTags",
        "tags": [
          "spaces"
        ],
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "type": "string"
            },
            "in": "path",
            "name": "namespaceSlug",
            "required": true
          },
          {
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "type": "string"
            },
            "in": "path",
            "name": "spaceSlug",
            "required": true
          }
        ],
        "security": [
          {},
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "name",
                      "commitId"
                    ],
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "commitId": {
                        "pattern": "^[0-9a-f]{40,64}$",
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "session": []
    }
  ]
}